Fix argument type for undocumented API "_setWindowNumber:". b=518546 r=smichaud
UltraBlame original commit: cb9c62cc27dcac1ab2b4ac0b185c69c4fe61d1da

+++ widget/src/cocoa/nsCocoaWindow.h
@@ -66,7 +66,7 @@ typedef struct _nsCocoaWindowList {
 
 
 
-- (void)_setWindowNumber:(int)aNumber;
+- (void)_setWindowNumber:(NSInteger)aNumber;
